Download, standardize, and merge XGR annotations

Merges a list of XGR annotations into a single GRangesList (or merged GRanges) object.

Usage

XGR_query(
  lib.selections = c("ENCODE_TFBS_ClusteredV3_CellTypes", "TFBS_Conserved",
    "Uniform_TFBS"),
  as_grangesList = FALSE,
  dat = NULL,
  n_top = NULL,
  nThread = 1
)

Arguments

lib.selections

Which XGR annotations to check overlap with. For full list of libraries see here (XGR on CRAN). Passed to the RData.customised argument in XGR::xRDataLoader.

as_grangesList

Return as a GRangesList, instead of a single merged GRanges object.

dat

data.table of genomic coordinates to query with. Set as NULL to return genome-wide data.

n_top

Filter to only the top N annotations that have the greatest amount of overlap with the genomic coordinates of dat.

nThread

Number of cores to parallelise across.

Value

GRangesList

See also

Examples

if (FALSE) { # \dontrun{
gr.lib <- echoannot::XGR_query(
    lib.selections = c("ENCODE_DNaseI_ClusteredV3_CellTypes"),
    dat = echodata::BST1,
    n_top = 1)
} # }